Option -m now limits output buffer.
This allows us to simulate 'short-stroking'
the encoder to hit a specific maximum buffer
size on the decode size.

Note that at this point the longest prefix
and the buffer needed are offset by +1,
which is annoying. Will change to tracking
buffer size explicitly, instead of prefix len.

Extra byte is needed in decoder for when we're
run into an unknown code and need to add the
missing character to the prefix.