Correct length-checks when decompressing known symbol.
Before we assumed the worst case: that the symbol was unknown
and _would_ be output to the destination buffer. This pessimistic
approach unfortunately meant that it was all but impossible to
decompress directly into a target buffer via the typical pattern
of adding bytes written to the destination pointer, while subtracting
them from the size left.

We would end up with a situation where we had say two more bytes to decode,
but the check would require three, causing decompression to stop
prematurely and then return LZW_DESTINATION_TOO_SMALL.

This wouldn't be an issue when decompressing to the same-sized buffer
as is done in the example code, but we _should_ support the pattern
which doesn't require unneccesary copying, when possible.