Initialize `symbol` to zero.
The GCC -fanalyzer warns about use of uninitialized value for
`symbol` in the output of new codes.

However, this analysis assumes that `prefix_len` can be zero,
which it would never be under normal (non-cosmic-ray-induced)
circumstances.

Nevertheless, initialize it to zero, and add an assert to
the fact that `prefix_len` is always greater than zero.